move characters, movies, spells and potions to /search; cleanup code
