+++ frontend/src/components/layout/Footer.js
@@ -1,6 +1,7 @@
 import { AppBar, Box, Container, Grid, Typography } from "@mui/material";
-import { getApiUrl, getGithubUrl } from "../../lib/utils";
+
 import Link from "../Link";
+import { getApiUrl, getGithubUrl } from "../../lib/utils";
 
 const Footer = () => {
   const footerData = [
@@ -30,19 +31,19 @@ const Footer = () => {
       content: [
         {
           name: "Characters",
-          link: "/characters",
+          link: "/search/characters",
         },
         {
           name: "Movies",
-          link: "/movies",
+          link: "/search/movies",
         },
         {
           name: "Potions",
-          link: "/potions",
+          link: "/search/potions",
         },
         {
           name: "Spells",
-          link: "/spells",
+          link: "/search/spells",
         },
       ],
     },
@@ -62,7 +63,13 @@ const Footer = () => {
   ];
 
   return (
-    <AppBar component="footer" color="transparent" elevation={0} position="static" sx={{ mt: 5, mb: 2 }}>
+    <AppBar
+      component="footer"
+      color="transparent"
+      elevation={0}
+      position="static"
+      sx={{ mt: 5, mb: 2 }}
+    >
       <Container disableGutters>
         <Grid container spacing={5}>
           {footerData.map((row) => (
@@ -79,7 +86,11 @@ const Footer = () => {
           ))}
         </Grid>
 
-        <Box style={{ textAlign: "center" }} pt={{ xs: 3, sm: 5 }} pb={{ xs: 4, sm: 0 }}>
+        <Box
+          style={{ textAlign: "center" }}
+          pt={{ xs: 3, sm: 5 }}
+          pb={{ xs: 4, sm: 0 }}
+        >
           Copyright &copy; Potter DB {new Date().getFullYear()}
         </Box>
       </Container>
