+++ frontend/src/components/ui/DatabaseAlert.js
@@ -3,8 +3,8 @@ import { useEffect, useRef, useState } from "react";
 import { Alert, AlertTitle, IconButton, Snackbar } from "@mui/material";
 import { Close } from "@mui/icons-material";
 
-import { getDatabaseStatus } from "../../lib/utils";
 import Link from "../Link";
+import { getDatabaseStatus } from "../../lib/utils";
 
 const DatabaseAlert = () => {
   const [databaseStatus, setDatabaseStatus] = useState(200);
