+++ frontend/src/pages/index.js
@@ -1,6 +1,6 @@
-import Meta from "../components/Meta";
-import LandingPageBanner from "../components/pages/index/LandingPageBanner";
 import FaqCard from "../components/pages/index/FaqCard";
+import LandingPageBanner from "../components/pages/index/LandingPageBanner";
+import Meta from "../components/Meta";
 
 const HomePage = () => {
   return (
